Only instantiate face detector once
There is a lot of lag making the robot unresponsive. The obvious target
place to start optimising is in the image processing code. Some timing
has been added to track the effect of changes.

The first change that has been made was to only instantiate the face
detector once, during the initialisation of the processor. This reduced
average processing time from ~0.4s per frame to ~0.09s per frame.