Move line adding until after face detection
The segments lines were being drawn on the image prior to the image
being loaded into the face detector. This may have been messing with the
face detection. It makes more sense to add these after the image has
been loaded into the detector and so there is no chance of it
interfering with the detection.